Bugfixing in BillHibernate, DrinkRepository, DrinkRepositoryImpl

+++ src/main/java/pl/connectis/restaurant/domain/service/DrinkRepository.java
@@ -11,9 +11,9 @@ public interface DrinkRepository {
     Drink createDrink(
             String name,
             String description,
-            double price,
+            BigDecimal price,
             Boolean is_available,
-            int portion_ml
+            BigDecimal portion_ml
     );
 
     Optional<Drink> getDrink(Long id);
